+++ server.js
@@ -1,16 +1,43 @@
 require("dotenv").config();
 var express = require("express");
 var exphbs = require("express-handlebars");
+var path = require("path");
 
 var db = require("./models");
 
 var app = express();
 var PORT = process.env.PORT || 3000;
 
+app.set("views", path.join(__dirname, "views"));
+app.set("view engine", "handlebars");
+
 // Middleware
-app.use(express.urlencoded({ extended: false }));
+app.use(express.urlencoded({ extended: false })); // app.use(express.urlencoded({ extended: true }));
 app.use(express.json());
-app.use(express.static("public"));
+app.use(express.static("public")); // app.use('/static', express.static('public'));
+
+// Auth middleware
+app.use(
+  require("express-session")({
+    secret: process.env.APP_SECRET,
+    resave: true,
+    saveUninitialized: false
+  })
+);
+
+const { ExpressOIDC } = require("@okta/oidc-middleware");
+const oidc = new ExpressOIDC({
+  issuer: `${process.env.OKTA_ORG_URL}/oauth2/default`,
+  client_id: process.env.OKTA_CLIENT_ID,
+  client_secret: process.env.OKTA_CLIENT_SECRET,
+  redirect_uri: `${process.env.HOST_URL}/authorization-code/callback`,
+  scope: "openid profile"
+});
+
+app.use(oidc.router);
+
+// Registration page
+app.use("/register", require("./routes/register"));
 
 // Handlebars
 app.engine(
@@ -28,6 +55,23 @@ require("./routes/userApiRoutes.js")(app);
 require("./routes/habitApiRoutes.js")(app);
 
 var syncOptions = { force: false };
+
+// Logout route
+app.get("/logout", (req, res) => {
+  if (req.userContext) {
+    const idToken = req.userContext.tokens.id_token;
+    const to = encodeURI(process.env.HOST_URL);
+    const params = `id_token_hint=${idToken}&post_logout_redirect_uri=${to}`;
+    req.logout();
+    res.redirect(
+      `${process.env.OKTA_ORG_URL}/oauth2/default/v1/logout?${params}`
+    );
+  } else {
+    res.redirect("/");
+  }
+});
+
+app.use("/", require("./routes/index"));
 
 // If running a test, set syncOptions.force to true
 // clearing the `testdb`
